Run servers in foreground
Currently they are run with shell=True to allow for argument parsing and
shell templating. This also has the drawback that they now run in the
background where they can be unnoticed and create confusion. This
changes it to no longer use shell expressions but rather plain argument
lists and the python-native argument for setting environment variables.